pyproject.toml
[project]
name = "today-report"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.154.0",
    "pygithub>=2.5.0",
    "python-dotenv>=1.0.1",
]
